Create RewardForMiner.js
This code is about to create an opportunity for miner that's why they can get a reward by completing mining. Because without the reward system, no miner can get interest mine block. And to mine block need multiple transaction systems in a block and in this program use this multiple transactions processes for pending transactions.